correct race between OpenReader and Wait
The cmd.Wait can finish before the OpenReader has consumed all of the
command's output pipe reader. Since the OpenReader is a non-main
goroutine, it doesn't stop the process from exiting unless we give the
main goroutine a reason to wait.

This PR does that by adding a little channel to the mix.

Fixes #3